moved routes to src directory
